Revert HlslTestUtils.h and DxcTestUtils.h/.cpp refactoring, remove HlslTestUtils.cpp
These files were reverted back to commit 41650a4a
(the commit before the first refactoring change)
